ignore node_modules
npm recommends checking in node_modules for anything that gets
deployed. While this does deploy to gh-pages, it only requires
jquery.js and geopattern.min.js, both of which are source controlled
already. The node_modules directory is only needed during development
or when used as a library, so it does not need to be checked in.