minor changes
I removed `index.js` as it was just a proxy for `lib/index.js`. The `main` field of `package.json` defines the entry point of the module.

NPM runs scripts with `node_modules/.bin` as part of `$PATH`, so specifying the path of the executables of eslint & mocha is not necessary.

As `js/jquery.min.js` may quickly get outdated and it isn't particularly to get jQuery, I removed it.